refactor(strategy-exercise): split 694L monolith into focused modules (Related to #1884)
TDD-first: 14 tests (3 new suites + original 3) all green.

- StrategyExercise.tsx: thin orchestrator, 694L → 80L
- OrderingExercise.tsx: drag-and-drop ordering (181L)
- TraitInferenceExercise.tsx: character trait MCQ + rescue (186L)
- GuidedStepsExercise.tsx: multi-step free_text+select + rescue (268L)
- strategyExerciseLogic.ts: pure rescue context builders + completion calc (98L)
- StrategyHeader.tsx: shared header sub-component (23L)

Test coverage:
- ordering_submit_marks_correct_only_when_order_matches
- trait_wrong_answer_records_attempt_and_opens_rescue_context
- guided_steps_persists_answers_feedback_and_all_done
